### 2024-03-28 #6552 @editorialbot
[PRE REVIEW]: MWRpy: A Python package for processing microwave radiometer data
**Submitting author:** @tobiasmarke (<a href="http://orcid.org/0000-0001-7804-9056">Tobias Marke</a>)
**Repository:** https://github.com/actris-cloudnet/mwrpy
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.8.3
