+++ paper/paper.bib
@@ -4,7 +4,8 @@
     year = {2023},
     publisher = {Zenodo},
     journal = {Zenodo repository},
-    url = {https://zenodo.org/records/7973553}
+    url = {https://zenodo.org/records/7973553},
+    doi = {10.5281/zenodo.7973552},
 }
 
 @article{IllingworthEt2007,
@@ -90,7 +91,7 @@
     volume = {38},
     number = {3},
     pages = {},
-    doi = {https://doi.org/10.1029/2002RS002634},
+    doi = {10.1029/2002RS002634},
     url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2002RS002634},
     eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2002RS002634},
     year = {2003}
