### 2024-04-02 c3709c1
Make median calculation more robust
