### 2024-04-30 cd0d35f
Fix adding HKD
