### 2024-05-02 658ef77
Improve azimuth check in sun flag
