+++ paper/paper.bib
@@ -27,13 +27,13 @@
     URL = {https://doi.org/10.1175/BAMS-88-6-883},
 }
 
-@book{ACTRIS_handbook2018,
-    author = {Häme, S. and Saporano, G. and Kivekäs, N. and Kaukolehto, M. and Rodriguez, E.},
-    title = {ACTRIS Stakeholder Handbook 2018},
-    year = {2018},
-    isbn = {978-952-336-066-2},
-    publisher = {Painotalo Trinket Oy},
-    url = {https://www.actris.eu},
+@article{Laj2024,
+    author = {Laj, P., and Coauthors},
+    title = {Aerosol, Clouds and Trace Gases Research Infrastructure – ACTRIS, the European research infrastructure supporting atmospheric science},
+    journal = {Bulletin of the American Meteorological Society},
+    year = {2024},
+    doi = {10.1175/BAMS-D-23-0064.1},
+    url = {https://journals.ametsoc.org/view/journals/bams/aop/BAMS-D-23-0064.1/BAMS-D-23-0064.1.xml},
 }
 
 @article{Tukiainen2020,
