### 2024-05-08 #6552 @Subho07
Dear Dr. Zhao,

I am interested in reviewing the submission.

Best regards,
Subhadip

-------------------------------------
Subhadip Dey
Assistant Professor,
Department of Agricultural and Food Engineering,
Indian Institute of Technology Kharagpur,
Kharagpur, India - 721302
*Mail To:* ***@***.*** ***@***.***>*
*Contact:* 03222-284540
