### 2024-05-08 #6733 @mengqi-z
👋 @tobiasmarke, @Subho07, and @kvenkman, Welcome to the review thread for the paper. All communication regarding this submission will take place here.

Please start by reading the "Reviewer instructions & questions" in the first comment above.

Reviewers, please create your checklists outlining JOSS requirements. As you assess the submission, mark any items you believe have been satisfied. Additionally, refer to the JOSS reviewer guidelines linked at the top of this thread.

Our aim is to collaborate with authors to help them meet our criteria. Reviewers are encouraged to submit issues and pull requests directly on the software repository. When doing so, please tag https://github.com/openjournals/joss-reviews/issues/6733 in the issue to create a link to this thread, enabling easy tracking. Please feel free to post comments, questions, and suggestions as they arise, rather than waiting until the entire package is reviewed.

We target completing reviews within 4-6 weeks, but please initiate your review well in advance. JOSS reviews are iterative, and your early feedback will help us stay on schedule.
