+++ mwrpy/level1/lev1_meta_nc.py
@@ -159,7 +159,7 @@ ATTRIBUTES_1B01 = {
         units="K",
     ),
     "azimuth_angle": MetaData(
-        long_name="Sensor azimuth angle",
+        long_name="Azimuth angle",
         standard_name="sensor_azimuth_angle",
         units="degree",
         comment="0=North, 90=East, 180=South, 270=West",
