+++ docs/source/fileformat.rst
@@ -51,7 +51,7 @@ MWR-1C01 file
 +---------------------------------------------+----------+-------------------------+------------------------------------------------------------------------------------------------------------------------------------------------------+-----------------------------------------------------------------+----------------------------------------+
 |ir_wavelength                                |float32   |ir_wavelength            |Wavelength of infrared channels                                                                                                                       |sensor_band_central_radiation_wavelength                         |m                                       |
 +---------------------------------------------+----------+-------------------------+------------------------------------------------------------------------------------------------------------------------------------------------------+-----------------------------------------------------------------+----------------------------------------+
-|azimuth_angle                                |float32   |time                     |Sensor azimuth angle                                                                                                                                  |sensor_azimuth_angle                                             |degree                                  |
+|azimuth_angle                                |float32   |time                     |Azimuth angle                                                                                                                                         |sensor_azimuth_angle                                             |degree                                  |
 +---------------------------------------------+----------+-------------------------+------------------------------------------------------------------------------------------------------------------------------------------------------+-----------------------------------------------------------------+----------------------------------------+
 |quality_flag                                 |int32     |time, frequency          |Quality flag                                                                                                                                          |                                                                 |1                                       |
 +---------------------------------------------+----------+-------------------------+------------------------------------------------------------------------------------------------------------------------------------------------------+-----------------------------------------------------------------+----------------------------------------+
@@ -132,7 +132,7 @@ MWR-Single file
 +---------------------------------------------+----------+-------------------------+------------------------------------------------------------------------------------------------------------------------------------------------------+-----------------------------------------------------------------+----------------------------------------+
 |temperature_quality_flag_status              |int32     |time                     |Quality flag status                                                                                                                                   |                                                                 |1                                       |
 +---------------------------------------------+----------+-------------------------+------------------------------------------------------------------------------------------------------------------------------------------------------+-----------------------------------------------------------------+----------------------------------------+
-|azimuth_angle                                |float32   |time                     |Sensor azimuth angle                                                                                                                                  |sensor_azimuth_angle                                             |degree                                  |
+|azimuth_angle                                |float32   |time                     |Azimuth angle                                                                                                                                  |sensor_azimuth_angle                                             |degree                                  |
 +---------------------------------------------+----------+-------------------------+------------------------------------------------------------------------------------------------------------------------------------------------------+-----------------------------------------------------------------+----------------------------------------+
 |height                                       |float32   |height                   |Height above mean sea level                                                                                                                           |height_above_mean_sea_level                                      |m                                       |
 +---------------------------------------------+----------+-------------------------+------------------------------------------------------------------------------------------------------------------------------------------------------+-----------------------------------------------------------------+----------------------------------------+
@@ -214,7 +214,7 @@ MWR-Multi file
 +---------------------------------------------+----------+-------------------------+------------------------------------------------------------------------------------------------------------------------------------------------------+-----------------------------------------------------------------+----------------------------------------+
 |time_bnds                                    |int32     |time, bnds               |Start and end time (UTC) of the measurements                                                                                                          |                                                                 |seconds since 1970-01-01 00:00:00.000   |
 +---------------------------------------------+----------+-------------------------+------------------------------------------------------------------------------------------------------------------------------------------------------+-----------------------------------------------------------------+----------------------------------------+
-|azimuth_angle                                |float32   |time                     |Sensor azimuth angle                                                                                                                                  |sensor_azimuth_angle                                             |degree                                  |
+|azimuth_angle                                |float32   |time                     |Azimuth angle                                                                                                                                  |sensor_azimuth_angle                                             |degree                                  |
[…]
