+++ mwrpy/version.py
@@ -1,4 +1,4 @@
 MAJOR = 0
-MINOR = 8
-PATCH = 3
+MINOR = 9
+PATCH = 0
 __version__ = f"{MAJOR}.{MINOR}.{PATCH}"
