### 2024-05-22 #6733 @Subho07
@mengqi-z @editorialbot  I think a exisitng program is not mentioned, for example, PyRTlib. This particular module works with radiative transfer functions which may also include the radiometer data. What is the difference between PyRTlib and MWRpy? Can the processes not be solved by PyRTlib? A flow chart may be required in the manuscript to understand the connectivity among the modules used in the software. Can the algorithms be applied for all type of microwave frequencies or any other pre-processing is required for some particular microwave frequencies?
 "is able to handle raw data from HATPRO manufactured by Radiometer Physics GmbH ": it sounds like the program is very specific to the product only, but can not be used for other products.
Otherwise the manuscript is well-written and easy to follow. The program also provides nice information on the installation procedure and has a test dataset that can be used to generate the test result for validation.
