+++ mwrpy/process_mwrpy.py
@@ -4,6 +4,8 @@ import datetime
 import logging
 import os
 
+import matplotlib.pyplot as plt
+
 from mwrpy.level1.write_lev1_nc import lev1_to_nc
 from mwrpy.level2.lev2_collocated import generate_lev2_multi, generate_lev2_single
 from mwrpy.level2.write_lev2_nc import lev2_to_nc
@@ -62,6 +64,7 @@ def main(args):
                 plot_product(product, date, args.site)
             except TypeError as err:
                 logging.error(err)
+            plt.close()
 
 
 def process_product(prod: str, date: datetime.date, site: str):
