+++ paper/paper.md
@@ -71,23 +71,26 @@ advancement of the IDL based processing software mwr_pro [@mwr_pro] and is able 
 manufactured by Radiometer Physics GmbH (RPG, https://www.radiometer-physics.de/), which is so far the only
 instrument type in the network. The output format, including metadata information, variable names, and file
 naming is designed to be compliant with the data structure and naming convention developed together with the
-EUMETNET Profiling Programme E-PROFILE [@Rüfenacht2021]. In this way,
-[MWRpy](https://actris-cloudnet.github.io/mwrpy/index.html#) improves data compatibility and fosters cross network
-collaborations. The processing chain is replacing the mode of operation in Cloudnet, which previously relied on
-pre-processed and non-harmonized MWR data, and therefore contributes to more ACTRIS data consistency. Statistical
-analysis of these consistent long-term data sets is expected to be beneficial not only for atmospheric studies, but
-also for improving knowledge on instrument operation and maintenance by monitoring key parameters from the
-instrument and mandatory regular absolute calibrations (approximately every 6 months). Future developments include
-the support of further instrument types, if present in the network. Furthermore, the flexible design of the code
-enables updating the retrievals of meteorological variables, which will be derived from a common approach.
+EUMETNET Profiling Programme E-PROFILE [@Rüfenacht2021], which is establishing a MWR network with the focus on
+near-real-time data provision[^1]. In this way, [MWRpy](https://actris-cloudnet.github.io/mwrpy/index.html#)
+improves data compatibility and fosters cross network collaborations. The processing chain is replacing the mode of
+operation in Cloudnet, which previously relied on pre-processed and non-harmonized MWR data, and therefore
+contributes to more ACTRIS data consistency. Statistical analysis of these consistent long-term data sets is
+expected to be beneficial not only for atmospheric studies, but also for improving knowledge on instrument operation
+and maintenance by monitoring key parameters from the instrument and mandatory regular absolute calibrations
+(approximately every 6 months). Future developments include the support of further instrument types, if present in
+the network. Furthermore, the flexible design of the code enables updating the retrievals of meteorological
+variables, which will be derived from a common approach.
+
+[^1]: E-PROFILE developed code for MWR processing (mwr_raw2l1, mwr_l12l2) can be found at https://github.com/MeteoSwiss
 
 # Code design
 
 [MWRpy](https://actris-cloudnet.github.io/mwrpy/index.html#) is designed to be used as a stand-alone software since
 it covers the full processing and visualization chain from raw data to higher level products, but it is also
 embedded in the Python implementation of the Cloudnet processing scheme CloudnetPy [@Tukiainen2020]. At first, data
-quality control is performed on the mandatory data fields of measured $T_B$ and instrument specific
-housekeeping data to generate quality flags. In a next step auxiliary data (e.g. from a weather station) are
+quality control is performed on the mandatory data fields of measured $T_B$ at various frequencies and instrument
+specific housekeeping data to generate quality flags. In a next step auxiliary data (e.g. from a weather station) are
 combined to produce daily netCDF files. Subsequently advanced meteorological variables are derived by applying
 retrieval coefficients and stored as separate daily files for variables originating from elevation scans (e.g.
 temperature profiles) and all remaining measuring modes (including vertical stare for e.g. LWP). Within the Cloudnet
@@ -95,6 +98,9 @@ processing framework the output of [MWRpy](https://actris-cloudnet.github.io/mwr
 and utilized by CloudnetPy, together with data streams from other ACTRIS-CCRES instruments, like cloud radar, to
 derive synergy products. All files, including calibration and retrieval information, and corresponding
[…]
