### 2024-05-23 02fb094
Add support for off-zenith SPC/INS
