### 2024-05-23 #6733 @kvenkman
Hi @tobiasmarke, could you clarify where the input MWR data path needs to be specified when running the command line tool `mwrpy/cli.py`? Could you add this to the documentation page, if it isn't already explicitly mentioned somewhere?

Also, following up on your previous comment, perhaps the paper would benefit from a brief mention of `PyRTlib` and how it's a simulation tool, compared to the data processing step that `MWRpy` serves.

The paper and software package look great otherwise 👍
