### 2024-05-24 16c7a51
Reference to radiative transfer model
