+++ paper/paper.bib
@@ -109,3 +109,15 @@
     url = {https://amt.copernicus.org/articles/5/1121/2012/},
     doi = {10.5194/amt-5-1121-2012}
 }
+
+@article{Larosa2024,
+    author = {Larosa, S. and Cimini, D. and Gallucci, D. and Nilo, S. T. and Romano, F.},
+    title = {PyRTlib: an educational Python-based library for non-scattering atmospheric microwave radiative transfer computations},
+    journal = {Geoscientific Model Development},
+    year = {2024},
+    volume = {17},
+    number = {5},
+    pages = {2053--2076},
+    url = {https://gmd.copernicus.org/articles/17/2053/2024/},
+    doi = {10.5194/gmd-17-2053-2024},
+}
