### 2024-05-24 d3ad319
Add information to command line usage
