### 2024-05-24 #6733 @tobiasmarke
Hi @kvenkman , thanks for reviewing.

The input data path can be specified in the configuration file `mwrpy/site_config/hatpro.yaml`. This is mentioned in the README under "Configuration" (including a link to an example file).
In addition, the documentation includes examples on how to run single modules and manually setting the data path (https://actris-cloudnet.github.io/mwrpy/mwrpy_processing.html). The information about the configuration file was added to the command line usage section (currently in the "docs" branch, but will be merged soon).

PyRTlib is now mentioned and referenced as an example for microwave radiative transfer calculations. Those will be needed for the retrieval development in the ACTRIS network.
