### 2024-05-29 #6733 @tobiasmarke
@Subho07 I've added more details on the E-PROFILE package and how it differs from MWRpy. In addition, an extended description about the potential use of PyRTlib as a forward simulator for retrieval development in the network is now included.
