+++ paper/paper.md
@@ -35,15 +35,15 @@ bibliography: paper.bib
 # Summary
 
 Ground-based passive microwave radiometers (MWRs) are deployed to obtain information on the vertical structure of
-temperature and water vapor mostly in the lower troposphere. In addition, they are used to derive the total column
-integrated liquid water content of the atmosphere, referred to as liquid water path (LWP). MWRs measure radiances,
+temperature and water vapor mostly in the lower troposphere. In addition, they are used to derive the total 
+column-integrated liquid water content of the atmosphere, referred to as liquid water path (LWP). MWRs measure radiances,
 given as brightness temperatures ($T_B$), typically in two frequency ranges along absorption features of water vapor
 and oxygen, as well as in window regions where the observations are sensitive to liquid water clouds. Profiles of
 temperature and humidity are retrieved together with the vertically integrated water vapor content (IWV) and LWP
-(e.g. @Crewell2003, @Löhnert2012). A specific elevation scanning configuration allows for an improved resolution for
+(e.g., @Crewell2003, @Löhnert2012). A specific elevation scanning configuration allows for an improved resolution for
 temperature profiles in the atmospheric boundary-layer [@Crewell2007]. The instruments can be operated continuously
-and provide temporally highly resolved observations of up to 1$s$, which make them a valuable tool for improving
-numerical weather forecast and climate models by studying the atmospheric water cycle, including cloud dynamics
+and provide temporally highly resolved observations of up to 1$s$, which makes them a valuable tool for improving
+numerical weather forecasts and climate models by studying the atmospheric water cycle, including cloud dynamics
 [@Westwater2004].
 
 One widely used application exploiting MWR data is the synergistic algorithm Cloudnet [@IllingworthEt2007], which
@@ -56,8 +56,8 @@ for MWRs in the network is hosted within ACTRIS Germany (ACTRIS-D).
 
 The European cloud remote sensing network will encompass around 30 stations, including mobile platforms, and covering
 different climatological zones. This network configuration enables investigations of similarities of atmospheric
-processes and long-term trends between those sites. Some of the participating stations have been operational already
-for more than a decade and Cloudnet products were derived based on their individual setup and processing
+processes and long-term trends between those sites. Some of the participating stations have been operational
+for more than a decade and Cloudnet products were derived based on their individual setups and processing
 algorithms. To ensure that the generated data sets are comparable, station operators are required to follow the
 ACTRIS-CCRES standard operating procedures and send raw data files to the central cloud remote sensing data center unit
 (CLU, http://cloudnet.fmi.fi). CLU provides data storage and provision, but also the centralized processing,
@@ -65,23 +65,23 @@ including visualization, in order to harmonize the data streams.
 
 # Statement of need
 
-[MWRpy](https://actris-cloudnet.github.io/mwrpy/index.html#) addresses the needs of a centralized processing,
+[`MWRpy`](https://actris-cloudnet.github.io/mwrpy/index.html#) addresses the needs of a centralized processing,
 quality control of MWR raw data, and deriving standardized output of meteorological variables. The Python code is an
-advancement of the IDL based processing software mwr_pro [@mwr_pro] and is able to handle raw data from HATPRO
+advancement of the Interactive Data Language (IDL)-based processing software mwr_pro [@mwr_pro] and is able to handle raw data from HATPRO
 manufactured by Radiometer Physics GmbH (RPG, https://www.radiometer-physics.de/), which is so far the only
 instrument type in the network. The output format, including metadata information, variable names, and file
[…]
