### 2024-06-12 bd3c7ce
Release version 1.0.0
