+++ mwrpy/version.py
@@ -1,4 +1,4 @@
-MAJOR = 0
-MINOR = 9
+MAJOR = 1
+MINOR = 0
 PATCH = 0
 __version__ = f"{MAJOR}.{MINOR}.{PATCH}"
