+++ README.md
@@ -2,6 +2,7 @@
 
 [![MWRpy tests](https://github.com/actris-cloudnet/mwrpy/actions/workflows/test.yml/badge.svg)](https://github.com/actris-cloudnet/mwrpy/actions/workflows/test.yml)
 [![PyPI version](https://badge.fury.io/py/mwrpy.svg)](https://badge.fury.io/py/mwrpy)
+[![DOI](https://zenodo.org/badge/619528669.svg)](https://zenodo.org/doi/10.5281/zenodo.11614185)
 
 MWRpy is a Python software to process RPG Microwave Radiometer data and is developed
 at the University of Cologne, Germany as part
