### 2024-06-14 #6733 @kthyng
@tobiasmarke  Paper looks good except please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example, "european" is not capitalized, but check for others.
