### 2024-06-20 #6733 @kthyng
Congratulations on your new publication @tobiasmarke! Many thanks to @mengqi-z and to reviewers @Subho07 and @kvenkman for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
