### 2024-06-20 #6733 @kthyng
@tobiasmarke If you're interested in reviewing for JOSS in the future, please register at https://reviewers.joss.theoj.org/.
